Events system (#4)
* preparing events base

* add simulation events

separatw input and simulation

* delete old input system(wip API macros)

* text enter reimplemetnation

* mouse coord input update step_1

* mouse coord input update step_2

* fix variables names

* camera movement update step_1

* camera movement update step_2

* Update input_handler.cpp(add some comments)

+++ Turret_1/code/map_structures/base_engine/coords/transforms.cpp
@@ -30,7 +30,7 @@ int t1::be::pixel(int tileCoord) {
 }
 
 float t1::be::pixelF(int tileCoord) {
-	return tileCoord * _TILE_ + _HALF_TILE_;
+	return static_cast<float>(tileCoord * _TILE_ + _HALF_TILE_);
 }
 
 PixelCoord t1::be::pixel(int tileCoordX, int tileCoordY) {
